+++ Panels/Header.js
@@ -5,8 +5,10 @@ import {getDirection, Spotlight} from '@enact/spotlight';
 import {getLastPointerPosition, hasPointerMoved} from '@enact/spotlight/src/pointer';
 import {getTargetByDirectionFromPosition} from '@enact/spotlight/src/target';
 import {Row, Cell} from '@enact/ui/Layout';
+import {useMeasurable} from '@enact/ui/Measurable';
 import Slottable from '@enact/ui/Slottable';
 import Toggleable from '@enact/ui/Toggleable';
+import {unit} from '@enact/ui/resolution';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
 import React from 'react';
@@ -18,7 +20,7 @@ import Skinnable from '../Skinnable';
 import {useScrollPosition} from '../useScroll/useScrollPosition';
 import WindowEventable from '../internal/WindowEventable';
 
-import {PanelsStateContext} from './Viewport';
+import {PanelsStateContext} from '../internal/Panels';
 
 import componentCss from './Header.module.less';
 
@@ -288,6 +290,18 @@ const HeaderBase = kind({
 		slotAfter: PropTypes.node,
 
 		/**
+		 * The method which receives the reference node to the slotAfter element, used to determine
+		 * the `slotSize`.
+		 *
+		 * @type {Function|Object}
+		 * @private
+		 */
+		slotAfterRef: PropTypes.oneOfType([
+			PropTypes.func,
+			PropTypes.shape({current: PropTypes.any})
+		]),
+
+		/**
 		 * A location for arbitrary elements to be placed to the left the title in LTR locales and
 		 * to the right in RTL locales
 		 *
@@ -305,6 +319,27 @@ const HeaderBase = kind({
 		 * @public
 		 */
 		slotBefore: PropTypes.node,
+
+		/**
+		 * The method which receives the reference node to the slotBefore element, used to determine
+		 * the `slotSize`.
+		 *
+		 * @type {Function|Object}
+		 * @private
+		 */
+		slotBeforeRef: PropTypes.oneOfType([
+			PropTypes.func,
+			PropTypes.shape({current: PropTypes.any})
+		]),
+
+		/**
+		 * The size for slotBefore and slotAfter.
+		 * This size is set by HeaderMeasurementDecorator for consistent title centering.
+		 *
+		 * @type {String}
+		 * @private
+		 */
+		slotSize: PropTypes.string,
 
 		/**
 		 * Text displayed below the title.
@@ -399,13 +434,13 @@ const HeaderBase = kind({
 	computed: {
 		backButtonAriaLabel: preferPropOverContext('backButtonAriaLabel'),
 		backButtonBackgroundOpacity: preferPropOverContext('backButtonBackgroundOpacity'),
-		className: ({backButtonAvailable, featureContent, hover, noBackButton, entering, centered, children, slotAbove, type, styler}) => styler.append(
+		className: ({backButtonAvailable, featureContent, hover, noBackButton, entering, centered, children, type, styler}) => styler.append(
 			{
 				featureContent,
 				centered,
 				// This likely doesn't need to be as verbose as it is, with the first 2 conditionals
 				showBack: (backButtonAvailable && !noBackButton && (hover || entering)),
-				withChildren: (Boolean(children) || Boolean(slotAbove))
+				withChildren: Boolean(children)
 			},
 			type
 		),
@@ -439,17 +474,20 @@ const HeaderBase = kind({
 		onClose,
 		slotAbove,
 		slotAfter,
+		slotAfterRef,
 		slotBefore,
+		slotBeforeRef,
+		slotSize,
 		subtitle,
 		title,
 		titleRef,
-		type,
 		...rest
 	}) => {
-		delete rest.featureContent;
 		delete rest.entering;
+		delete rest.featureContent;
 		delete rest.onHideBack;
 		delete rest.onShowBack;
+		delete rest.type;
 
 		// Set up the back button
 		const backButton = (backButtonAvailable && !noBackButton ? (
@@ -478,17 +516,20 @@ const HeaderBase = kind({
 			/>
 		) : null);
 
-		// In wizard type, if one slot is filled, automatically include the other to keep the title balanced.
-		// DEV NOTE: Currently, the width of these is not synced, but can/should be in a future update.
-		const bothBeforeAndAfter = (type === 'wizard' && (slotAfter || slotBefore));
+		// Only provide the synced cell size if the title should be centered, beyond that case,
+		// the cell sizes don't need to be synced.
+		const syncCellSize = (centered ? slotSize : null);
 
+		// The side Cells are always present, even if empty, to support the measurement ref.
 		return (
 			<header {...rest}>
 				{slotAbove ? <nav className={css.slotAbove}>{slotAbove}</nav> : null}
 				<Row className={css.titlesRow} align="center" ref={titleRef}>
-					{(bothBeforeAndAfter || slotBefore || backButton) ? (
-						<Cell shrink className={css.slotBefore}>{backButton}{slotBefore}</Cell>
-					) : null}
+					<Cell className={css.slotBefore} shrink={!syncCellSize} size={syncCellSize}>
+						<span ref={slotBeforeRef} className={css.slotSizer}>
+							{backButton}{slotBefore}
+						</span>
+					</Cell>
 					<Cell className={css.titleCell}>
 						<Heading
 							aria-label={title}
@@ -512,9 +553,11 @@ const HeaderBase = kind({
 							{subtitle}
 						</Heading>
 					</Cell>
-					{(bothBeforeAndAfter || slotAfter || closeButton) ? (
-						<Cell shrink className={css.slotAfter}>{slotAfter}{closeButton}</Cell>
-					) : null}
+					<Cell className={css.slotAfter} shrink={!syncCellSize} size={syncCellSize}>
+						<span ref={slotAfterRef} className={css.slotSizer}>
+							{slotAfter}{closeButton}
+						</span>
+					</Cell>
 				</Row>
 				{children ? <nav className={css.slotBelow}>{children}</nav> : null}
 				{line}
@@ -530,10 +573,41 @@ const CollapsingHeaderDecorator = (Wrapped) => {
 	};
 };
 
+const HeaderMeasurementDecorator = (Wrapped) => {
+	return function HeaderMeasurementDecorator (props) { // eslint-disable-line no-shadow
+		const {ref: slotBeforeRef, measurement: {width: slotBeforeWidth = 0} = {}} = useMeasurable() || {};
+		const {ref: slotAfterRef, measurement: {width: slotAfterWidth = 0} = {}} = useMeasurable() || {};
+		const [{slotSize, prevSlotBeforeWidth, prevSlotAfterWidth}, setSlotSize] = React.useState({});
+
+		// If the slot width has changed, re-run this.
+		if (slotBeforeWidth !== prevSlotBeforeWidth || slotAfterWidth !== prevSlotAfterWidth) {
+			const largestSlotSize = Math.max(slotBeforeWidth, slotAfterWidth);
+
+			// And only do this the largest slot is a different value this time around.
+			if (slotSize !== largestSlotSize) {
+				setSlotSize({
+					slotSize: largestSlotSize,
+					prevSlotBeforeWidth: slotBeforeWidth,
+					prevSlotAfterWidth: slotAfterWidth
+				});
+			}
+		}
+
+		const measurableProps = {
+			slotBeforeRef,
+			slotAfterRef,
+			slotSize: unit(slotSize, 'rem')
+		};
+
+		return <Wrapped {...props} {...measurableProps} />;
+	};
+};
+
 const HeaderDecorator = compose(
 	Slottable({slots: ['title', 'subtitle', 'slotAbove', 'slotAfter', 'slotBefore']}),
 	Skinnable,
 	CollapsingHeaderDecorator,
+	HeaderMeasurementDecorator,
 	Toggleable({prop: 'hover', activate: 'onShowBack', deactivate: 'onHideBack', toggle: null}),
 	WindowEventable({globalNode: 'document', onKeyDown: handleWindowKeyPress})
 );