+++ samples/qa-a11y/src/App/Home.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Home = () => (
+	<div>
+		<p>
+			Enact Component patterns for accessibility features.
+		</p>
+	</div>
+);
+
+export default Home;