+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [1.0.0-beta.2] - 2020-05-04
+
+### Deprecated
+
+- `sandstone/Panels.WizardPanel` and `sandstone/Panels.View`, replaced with `sandstone/WizardPanels` and `sandstone/WizardPanels.Panel` respectively
+- `sandstone/Panels.FlexiblePopupPanels`, replaced with `sandstone/FlexiblePopupPanels`
+- `sandstone/Panels.FixedPopupPanels`, replaced with `sandstone/FixedPopupPanels`
+
+### Added
+
+- `sandstone/WizardPanels` props `nextButtonAriaLabel`, `prevButtonAriaLabel`, `noNextButton`, `noPrevButton`, and `noSteps`
+- `sandstone/Scroller` and `sandstone/VirtualList` prop `initialHiddenHeight` to provide the height of the vertical scrollbar when the `featureContent` prop in the panel is set to true
+- `sandstone/Input.InputPopup` component
+
+### Fixed
+
+- `sandstone/Header` centering
+- `sandstone/Input.InputField` disabled colors
+- `sandstone/WizardPanels` to hide previous and next buttons appropriately
+- `sandstone/TabLayout` to support disabled tabs
+
 ## [1.0.0-beta.1] - 2020-04-27
 
 ### Removed
@@ -29,6 +50,8 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Fixed
 
+- `sandstone/BodyText` font weight
+- `sandstone/BodyText` line-wrap and `noWrap` capabilities
 - `sandstone/DatePicker` and `sandstone/TimePicker` to match current designs
 - `sandstone/Dropdown` to focus on selected option
 - `sandstone/Picker` horizontal joined height in large text mode